Former-commit-id: 49b5dc19b6b339b5485336c8935eca93e44cfa64
Former-commit-id: 7857a5e57e5f38314b65e73d55c220cd170330cc

+++ application/router/EdgeCosts.cc
@@ -57,7 +57,7 @@ void EdgeCosts::insert(int d)
 //Returns the % of readings of a certain travel time
 double EdgeCosts::percentAt(int d)
 {
-    return (double)data[d] / (double)count;
+    return double(data[d]) / count;
 }
 
 }
